adapt content width to sidebar state
