+++ src/index.css
@@ -81,6 +81,12 @@ p {
 
 .fix-content-width {
   /* width: calc(100vw - 80px) */
+  margin-left: 250px;
+  width: 100%;
+}
+
+.fix-content-width-hidden {
+  /* width: calc(100vw - 80px) */
   margin-left: 60px;
   width: 100%;
 }
@@ -107,13 +113,6 @@ p {
   position: absolute;
   right: 0px;
   bottom: 0px;
-}
-
-@media only screen and (min-device-width: 768px) {
-  .fix-content-width {
-    /* width: calc(100vw - 270px) */
-    margin-left: 250px
-  }
 }
 
 @media only screen and (min-width: 992px) {
